Add a spec option to build a combined library
Upstream has an option to build a combined library but it'll clash
with symbol/library versioning once it lands. Use a linker script
instead of an actual ELF DSO to side-step the issue.

GROUP() is used instead of INPUT() to avoid the need for manual
ordering of the umphteen libraries.